#3 Multiple spaces - work in progress
