+++ app/views/spaces/spaces.erb
@@ -17,9 +17,9 @@
       <% @spaces.each do | space | %>
       <li>
         <%= space.name %>
-        <BR>
+        <br>
         <%= space.description %>
-        <BR>
+        <br>
         <%= space.price %>
       </li>
       <% end %>
