+++ spec/feature/listing_space_spec.rb
@@ -1,13 +1,7 @@
 feature 'Listing spaces' do
-    scenario 'user can list a space' do
-
-      visit '/spaces'
-      click_button 'List a Space'
-      expect(current_path).to eq('/spaces/new')
-      fill_in 'Name', with: 'My beautiful home'
-      fill_in 'Description', with: 'It has wifi'
-      fill_in 'Price', with: '100'
-      click_button 'List my Space'
+    scenario 'a signed in user can list a space' do
+      sign_up
+      list_a_space
       expect(current_path).to eq('/spaces')
 
       within 'ul#spaces' do
@@ -16,4 +10,18 @@ feature 'Listing spaces' do
         expect(page).to have_content('100')
       end
     end
+
+    scenario 'signed in user can list multiple spaces' do
+      sign_up
+      list_a_space
+      click_button 'List a Space'
+      fill_in 'Name', with: 'My ugly smelly home'
+      fill_in 'Description', with: 'It has no wifi but it has bugs'
+      fill_in 'Price', with: '1'
+      click_button 'List my Space'
+      user = User.first
+      expect(user.spaces.map(&:name)).to include("My ugly smelly home")
+      expect(user.spaces.map(&:name)).to include("My beautiful home")
+    end
+
 end
