+++ spec/feature/sign_up_spec.rb
@@ -9,12 +9,3 @@ feature 'User Sign Up' do
     expect { sign_up(password_confirmation: 'wrong') }.not_to change(User, :count)
   end
 end
-
-def sign_up(email: 'tim.chipperfield@gmail.com', password: 'christmastrees', password_confirmation: 'christmastrees' )
-    visit '/'
-    expect(page.status_code).to eq(200)
-    fill_in :email, with: email
-    fill_in :password, with: password
-    fill_in :password_confirmation, with: password_confirmation
-    click_button 'Sign Up'
-end
