+++ spec/spec_helper.rb
@@ -9,6 +9,7 @@ require 'capybara/rspec'
 require 'rspec'
 require './app/models/user'
 require 'database_cleaner'
+require 'web_helpers'
 
 
 
@@ -17,18 +18,18 @@ Capybara.app = MakersBnB
 RSpec.configure do |config|
   config.include Capybara::DSL
 
-  config.before(:suite) do
-    DatabaseCleaner.strategy = :transaction
-    DatabaseCleaner.clean_with(:transaction)
-  end
-
-  config.before(:each) do
-    DatabaseCleaner.start
-  end
-
-  config.after(:each) do
-    DatabaseCleaner.clean
-  end
+  # config.before(:suite) do
+  #   DatabaseCleaner.strategy = :transaction
+  #   DatabaseCleaner.clean_with(:transaction)
+  # end
+  #
+  # config.before(:each) do
+  #   DatabaseCleaner.start
+  # end
+  #
+  # config.after(:each) do
+  #   DatabaseCleaner.clean
+  # end
 
 
   config.expect_with :rspec do |expectations|
